shuffled each class into its own file/minor naming corrections
