+++ Mystic_Tuner_Backend/Database_connector/test_card_queries.py
@@ -1,5 +1,4 @@
-from database_queries import card_queries
-from database_connector import DatabaseConnector
+from card_queries import card_queries
 import pytest
