+++ Mystic_Tuner_Backend/Database_connector/test_database_connection.py
@@ -1,7 +1,7 @@
-from database_connector import DatabaseConnector
+from database_connector import database_connector
 import pytest
 
 def test_establish_connection():
-    instance = DatabaseConnector("localhost", "Mystic_Tuner_Application", "MT_Admin", "admin", 5433)
+    instance = database_connector("localhost", "Mystic_Tuner_Application", "MT_Admin", "admin", 5433)
     assert instance != None
